timers: rename to timers & buffs
