+++ runelite-client/src/main/java/net/runelite/client/plugins/timersandbuffs/GameIndicator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Jordan Atwood <jordan.atwood423@gmail.com>
+ * Copyright (c) 2018, Tyler <https://github.com/tylerthardy>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,10 +22,35 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.timers;
+package net.runelite.client.plugins.timersandbuffs;
 
-enum GameTimerImageType
+import java.awt.Color;
+import lombok.AccessLevel;
+import lombok.Getter;
+import net.runelite.api.SpriteID;
+
+@Getter(AccessLevel.PACKAGE)
+enum GameIndicator
 {
-	ITEM,
-	SPRITE
+	VENGEANCE_ACTIVE(SpriteID.SPELL_VENGEANCE_OTHER, GameTimerImageType.SPRITE, "Vengeance active");
+
+	private final String description;
+	private String text;
+	private Color textColor;
+	private final int imageId;
+	private final GameTimerImageType imageType;
+
+	GameIndicator(int imageId, GameTimerImageType idType, String description, String text, Color textColor)
+	{
+		this.imageId = imageId;
+		this.imageType = idType;
+		this.description = description;
+		this.text = text;
+		this.textColor = textColor;
+	}
+
+	GameIndicator(int imageId, GameTimerImageType idType, String description)
+	{
+		this(imageId, idType, description, "", null);
+	}
 }
